requirements.txt
opencensus 
opencensus-ext-azure
opencensus-ext-logging
python-dotenv